+++ src/user/index.js
@@ -1,4 +1,4 @@
-import { getCookie, setCookie } from "./cookie/index.js";
+import { getCookie, deleteCookie, setCookie } from "./cookie/index.js";
 import { getScrollbarWidth } from "./getScrollbarWidth/index.js";
 import { isMobileDevice } from "./isMobileDevice/index.js";
 import { isTouchDevice } from "./isTouchDevice/index.js";
@@ -9,4 +9,5 @@ export {
   isTouchDevice,
   setCookie,
   isMobileDevice,
+  deleteCookie,
 };