+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+### 1.0.5
+- deleteCookie/index.js: new util added;
+- getFromServer/index.js: `arrayBuffer` response type added;
+- API.md updated;
+- Development dependencies updated;
+
 ### 1.0.3
-- getAsyncMap.js: unnecessary try/catch removed;
+- getAsyncMap/index.js: unnecessary try/catch removed;
 - API.md created;
 
 ### 1.0.0